requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=2.5
Flask-Login>=0.6
Flask-Mail>=0.9
itsdangerous>=2.0
psycopg2-binary>=2.9
sendgrid>=6.0
gunicorn
python-dotenv>=0.19
requests>=2.25 